CONSOLE/TEST: add release handling for squash fs packages - add instance tests for 'eos squash' command set

+++ console/commands/com_squash.cc
@@ -79,10 +79,13 @@ com_squash(char* arg1)
       goto com_squash_usage;
     }
 
-    XrdOucString garbage = subtokenizer.GetToken();
-
-    if (garbage.length()) {
-      goto com_squash_usage;
+    if ( (cmd != "trim-release") && (cmd != "new-release") ){
+      XrdOucString garbage = subtokenizer.GetToken();
+      if (garbage.length()) {
+	goto com_squash_usage;
+      } else {
+	break;
+      }
     } else {
       break;
     }
@@ -137,19 +140,21 @@ com_squash(char* arg1)
 	  fulloption.endswith(".tar.gz")) {
 	int sub_rc = 0;
 	// squash rm
-	std::string subcommand = "rm ";
+	std::string subcommand = "rm \"";
 	subcommand += path.c_str();
+	subcommand += "\"";
 	com_squash((char*)subcommand.c_str());
 	sub_rc |= global_retc;
 	// squash new
-	subcommand = "new ";
+	subcommand = "new \"";
 	subcommand += path.c_str();
+	subcommand += "\"";
 	com_squash((char*)subcommand.c_str());
 	sub_rc |= global_retc;
 	// download
-	std::string shellcmd="cd "; 
+	std::string shellcmd="cd \"";
 	shellcmd += path.c_str();
-	shellcmd += ";";
+	shellcmd += "\";";
 	shellcmd += "curl ";
 	shellcmd += url;
 	shellcmd += " /dev/stdout | ";
@@ -162,8 +167,9 @@ com_squash(char* arg1)
 	  return (0);
 	}
 	// squash pack
-	subcommand = "pack ";
+	subcommand = "pack \"";
 	subcommand += path.c_str();
+	subcommand += "\"";
 	com_squash((char*)subcommand.c_str());
 	sub_rc |= global_retc;
 	if (sub_rc) {
@@ -188,6 +194,7 @@ com_squash(char* arg1)
     squashpack += ".sqsh";
     std::string shellcmd = "mksquashfs ";
     char linktarget[4096];
+    memset(linktarget,0, sizeof(linktarget));
     ssize_t rl;
 
     // resolve symlink
@@ -280,6 +287,12 @@ com_squash(char* arg1)
     squashpack += packagepath.GetName();
     squashpack += ".sqsh";
 
+    if (stat(squashpack.c_str(), &buf)) {
+      fprintf(stderr,"error: the squashfs package file is missing for this label!\n");
+      global_retc = ENOENT;
+      return (0);
+    }
+
     if (!lstat(packagepath.GetPath(), &buf)) {
       if (unlink(packagepath.GetPath())) {
         fprintf(stderr,
@@ -340,23 +353,25 @@ com_squash(char* arg1)
     mkpath += username;
     mkpath += "/eosxd/mksquash/";
 
-    if (option != "f") {
+    if (option != "f" ) {
       if (mounttarget.beginswith(mkpath.c_str())) {
 	fprintf(stderr, "error: squash image is already unpacked!\n");
 	global_retc = EINVAL;
 	return (0);
       }
       
-      // remove any mounts
-      std::string umountcmd = "umount -f -l ";
-      umountcmd += mounttarget.c_str();
-      system(umountcmd.c_str());
-      
-      if (rmdir(mounttarget.c_str())) {
-	if (errno != ENOENT) {
-	  fprintf(stderr,
+      if (!geteuid()) {
+	// remove any mounts - only possible as root
+	std::string umountcmd = "umount -f -l ";
+	umountcmd += mounttarget.c_str();
+	system(umountcmd.c_str());
+
+	if (rmdir(mounttarget.c_str())) {
+	  if (errno != ENOENT) {
+	    fprintf(stderr,
 		  "error: failed to unlink local mount directory path='%s' errno=%d\n",
-		  mounttarget.c_str(), errno);
+		    mounttarget.c_str(), errno);
+	  }
 	}
       }
     }
@@ -452,30 +467,25 @@ com_squash(char* arg1)
       linktarget[rl] = 0;
     }
 
-    if (stat(linktarget, &buf)) {
-      if (!S_ISLNK(buf.st_mode)) {
-        fprintf(stderr, "error: this does not look like a squash image smart link\n");
-      } else {
-        fprintf(stderr, "error: cannot find local package directory '%s'\n",
-                linktarget);
-      }
-
-      global_retc = EINVAL;
-      return (0);
-    }
-
     XrdOucString mounttarget = linktarget;
     std::string mkpath = "/var/tmp/";
     mkpath += username;
     mkpath += "/eosxd/mksquash/";
 
     if (mounttarget.beginswith(mkpath.c_str())) {
+      if (stat(linktarget, &buf)) {
+	fprintf(stderr, "error: cannot find local package directory '%s'\n",
+		linktarget);
+	global_retc = EINVAL;
+	return (0);
+      }
+
       fprintf(stderr,
               "info: squashfs image is currently unpacked/open for local RW mode - use 'eos squash pack %s' to close image\n",
               packagepath.GetPath());
     } else {
       fprintf(stderr,
-              "info: squashfs image is currently open in RO mode - use 'eos squash unpack %s' to open image locally\n",
+              "info: squashfs image is currently packed - use 'eos squash unpack %s' to open image locally\n",
               packagepath.GetPath());
     }
   }
@@ -515,6 +525,228 @@ com_squash(char* arg1)
     }
   }
 
+
+  if (cmd == "rm-release") {
+    std::string scmd = "info-release ";
+    scmd += path.c_str();
+    com_squash((char*) scmd.c_str());
+    if (!global_retc) {
+      fprintf(stderr,"info: wiping squashfs releases under '%s'\n", path.c_str());
+      eos::common::Path packagepath(path.c_str());
+      std::string nextrelease = std::string(packagepath.GetPath()) + std::string("/next");
+      std::string currentrelease = std::string(packagepath.GetPath()) + std::string("/current");
+      std::string archive = std::string(packagepath.GetPath()) + std::string("/.archive");
+
+      fprintf(stdout,"info: wiping links current,next ... \n");
+      ::unlink(currentrelease.c_str());
+      ::unlink(nextrelease.c_str());
+      if (archive.substr(0,5) == "/eos/") {
+	fprintf(stdout,"info: wiping archive ...\n");
+	scmd = "eos rm -rf "; scmd += archive;
+	std::string out = eos::common::StringConversion::StringFromShellCmd(scmd.c_str());
+	fprintf(stdout,"%s", out.c_str());
+      }
+
+      if (::rmdir(packagepath.GetPath())) {
+	fprintf(stderr,"error: failed to clean squashfs release under '%s'\n", path.c_str());
+	global_retc = errno;
+      }
+
+      return (0);
+    } else {
+      fprintf(stderr,"info: there is no squashfs release uner '%s'\n", path.c_str());
+      return (0);
+    }
+  }
+  if (cmd == "new-release") {
+    eos::common::Path packagepath(path.c_str());
+
+    XrdOucString version = subtokenizer.GetToken();
+
+    std::string packagename = packagepath.GetName();
+    std::string now = eos::common::StringConversion::StringFromShellCmd("date '+%Y%m%d%H%M%S'");
+    if (version.length()) {
+      // overwrite with the given version number
+      now = version.c_str();
+    }
+
+    now.erase(now.find_last_not_of(" \n\r\t")+1);
+
+    std::string archivepath = std::string(packagepath.GetPath()) + std::string("/") + std::string(".archive/");
+    std::string archivepackage = archivepath + packagename + std::string("-") + now;
+    std::string nextrelease = std::string(packagepath.GetPath()) + std::string("/next");
+
+    eos::common::Path archpath(archivepackage.c_str());
+    if (!archpath.MakeParentPath(S_IRWXU | S_IRGRP | S_IXGRP | S_IROTH | S_IXOTH)) {
+      fprintf(stderr,"error: couldn't create '%s'\n", archpath.GetParentPath());
+      global_retc = errno;
+      return (0);
+    }
+    // in the unlikely case this command was executed within 1 second twice ...
+    ::unlink(archivepackage.c_str());
+    ::unlink(nextrelease.c_str());
+    std::string scmd = "new \"";
+    scmd += archivepackage;
+    scmd += "\"";
+    int rc = com_squash((char*)scmd.c_str());
+    if (!rc) {
+      rc = ::symlink(archivepackage.c_str(), nextrelease.c_str());
+      if (rc) {
+	fprintf(stderr,"error: failed to create symbolic link for next release '%s'\n", nextrelease.c_str());
+	global_retc = errno;
+	return (0);
+      }
+    } else {
+      fprintf(stderr,"error: failed to create squash package for a new release\n");
+      return (0);
+    }
+
+    fprintf(stderr,"info: install the new release under '%s'\n", nextrelease.c_str());
+    return (0);
+  }
+
+  if (cmd == "pack-release") {
+    char lname[4096];
+    memset(lname,0, sizeof(lname));
+
+    eos::common::Path packagepath(path.c_str());
+    std::string nextrelease = std::string(packagepath.GetPath()) + std::string("/next");
+    std::string currentrelease = std::string(packagepath.GetPath()) + std::string("/current");
+    std::string hiddencurrentrelease = std::string(packagepath.GetPath()) + std::string("/.current");
+
+    if (::readlink(nextrelease.c_str(),lname, sizeof(lname)) < 0) {
+      fprintf(stderr,"error: failed to find an open release package under '%s'\n", nextrelease.c_str());
+      global_retc = errno;
+      return (0);
+    } else {
+      std::string scmd = "pack ";
+      scmd += "\"";
+      scmd += lname;
+      scmd += "\"";
+      int rc = com_squash((char*)scmd.c_str());
+      if (!rc) {
+	rc = ::unlink(nextrelease.c_str());
+	if (rc) {
+	  fprintf(stderr,"error: failed to unlink open release package under '%s'\n", nextrelease.c_str());
+	  global_retc = errno;
+	  return (0);
+	}
+
+	rc = ::symlink(lname, hiddencurrentrelease.c_str());
+	if (rc) {
+	  fprintf(stderr,"error: failed to symlink current release package under '%s'\n", hiddencurrentrelease.c_str());
+	  global_retc = errno;
+	  return (0);
+	} else {
+	  rc = ::rename(hiddencurrentrelease.c_str(), currentrelease.c_str());
+	  if (rc) {
+	    fprintf(stderr,"error: failed to move '%s' to '%s'\n", hiddencurrentrelease.c_str(), currentrelease.c_str());
+	    global_retc = errno;
+	    return (0);
+	  } else {
+	    fprintf(stdout,"info: new release available under '%s'\n", currentrelease.c_str());
+	    return (0);
+	  }
+	}
+      } else {
+	fprintf(stderr,"error: failed to pack squash package for a new release\n");
+	return (0);
+      }
+    }
+  }
+
+  if (cmd == "info-release") {
+    std::string scmd = "trim-release \"";
+    scmd += path.c_str();
+    scmd += "\" ";
+    scmd += "999999 999999";
+
+    com_squash((char*)scmd.c_str());
+    return (0);
+  }
+
+  if (cmd == "trim-release") {
+    eos::common::Path packagepath(path.c_str());
+    XrdOucString keepdays = subtokenizer.GetToken();
+    XrdOucString keepversions  = subtokenizer.GetToken();
+
+    std::string current = packagepath.GetPath();
+    std::string archive = packagepath.GetPath();
+    current += "/current";
+    archive += "/.archive";
+    struct stat buf;
+    if (::lstat(current.c_str(), &buf)) {
+      fprintf(stderr,"error: I cannot find any current release under '%s'\n", current.c_str());
+      global_retc = EINVAL;
+      return (0);
+    }
+    if (::lstat(archive.c_str(), &buf)) {
+      fprintf(stderr,"error: I cannot find any archive release under '%s'\n", archive.c_str());
+      global_retc = EINVAL;
+      return (0);
+    }
+
+    if (!keepdays.length()) {
+      fprintf(stderr,"error: you have to specify the number of days you want to keep releases : squash trim-release <path> <n-days> [<max-versions]\n");
+      global_retc = EINVAL;
+      return (0);
+    }
+
+    size_t n_keepdays = strtol(keepdays.c_str(),0,10);
+
+    if (!n_keepdays) {
+      fprintf(stderr,"error: you have to specify the number of days you want to keep releases : squash trim-release <path> <n-days>\n");
+      global_retc = EINVAL;
+      return (0);
+    }
+
+    size_t n_keepversions = keepversions.length()?strtol(keepversions.c_str(),0,10):0;
+
+    if (!n_keepversions) {
+      fprintf(stderr,"info: no !=0 version limit specified ...\n");
+      keepversions = "1000000";
+    } else {
+      // we have to pass keepversions + 1 to the find commands
+      keepversions = std::to_string(n_keepversions+1).c_str();
+    }
+
+    std::string find1,find2,find3,find4,find5;
+
+    find1 = find2 = find3 = find4 = find5 = "find ";
+    find1 += packagepath.GetPath();
+    find2 += packagepath.GetPath();
+    find3 += packagepath.GetPath();
+    find4 += packagepath.GetPath();
+    find5 += packagepath.GetPath();
+
+    find1 += " -type f -mtime +";
+    find2 += " -type l -mtime +";
+    find1 += keepdays.c_str();
+    find2 += keepdays.c_str();
+    find1 += " -delete";
+    find2 += " -delete";
+    find3 += "/.archive/ -type f -printf '%Ts\t%h/%f\n'     | sort -rn | tail -n +";
+    find4 += "/.archive/ -type l -printf '%Ts\t%h/%f\n'     | sort -rn | tail -n +";
+    find3 += keepversions.c_str();
+    find4 += keepversions.c_str();
+    find3 += " | cut -f2- | xargs -r rm";
+    find4 += " | cut -f2- | xargs -r rm";
+    find5 += " -type l";
+
+    eos::common::StringConversion::StringFromShellCmd(find1.c_str());
+    eos::common::StringConversion::StringFromShellCmd(find2.c_str());
+    eos::common::StringConversion::StringFromShellCmd(find3.c_str());
+    eos::common::StringConversion::StringFromShellCmd(find4.c_str());
+    std::string out = eos::common::StringConversion::StringFromShellCmd(find5.c_str());
+
+    fprintf(stdout,"---------------------------------------------------------------------------\n");
+    fprintf(stdout,"- releases of '%s' \n", packagepath.GetPath());
+    fprintf(stdout,"---------------------------------------------------------------------------\n");
+    fprintf(stdout,"%s", out.c_str());
+    fprintf(stdout,"---------------------------------------------------------------------------\n");
+    return (0);
+  }
+
   if (!ok) {
     goto com_squash_usage;
   }
@@ -546,14 +778,34 @@ com_squash_usage:
 	  "       squash relabel <path>                                              : relable a squashfs image link e.g. after an image move in the namespace\n");
   fprintf(stdout,"\n");
 
-  fprintf(stdout, 
+  /*  fprintf(stdout,
 	  "       squash roll <path>                                                 : will create a squash package from the EOS directory pointed by <path\n");
   fprintf(stdout,"\n");
   fprintf(stdout, 
 	  "       squash unroll <path>                                               : will store the squash package contents unpacked into the EOS package directory\n");
   fprintf(stdout,"\n");
+  */
+
   fprintf(stdout, 
 	  "       squash install --curl=https://<package>.tgz|.tar.gz <path>         : create a squashfs package from a web archive under <path>\n");
+
+  fprintf(stdout,
+          "       squash new-release <path> [<version>]                                : create a new squashfs release under <path> - by default versions are made from timestamp, but this can be overwritten using the version field\n");
+
+  fprintf(stdout,
+          "       squash pack-release <path>                                         : pack a squashfs release under <path>\n");
+
+
+  fprintf(stdout,
+          "       squash info-release <path>                                         : show all release revisions under <path> <path>\n");
+
+
+  fprintf(stdout,
+          "       squash trim-release <path> <keep-days> [<keep-versions>]           : trim  releases older than <keep-days> and keep maximum <keep-versions> of release\n");
+
+  fprintf(stdout,
+          "       squash rm-release <path>                                           : delete all squahfs releases udner <path>\n");
+
   global_retc = EINVAL;
   return (0);
 }
